requirements.txt
# 旅行代理程序依赖包
python-dotenv>=1.0.0
openai>=1.0.0
